changed the name of the react element file to react-comp and vice versa
