+++ src/components/react-element.js
@@ -1,8 +1,21 @@
-import React, {Component} from 'react';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ReactElement from './react-comp';
+import {noView, inject, customElement} from 'aurelia-framework';
 
-export default class HelloWorld extends Component {
+@noView()
+@inject(Element)
+@customElement('react-component')
+export class ReactComponent {
+  constructor(element) {
+    this.element = element;
+  }
 
   render() {
-    return <h1>Hello World</h1>;
+    ReactDOM.render(<ReactElement/>, this.element);
+  }
+
+  bind() {
+    this.render();
   }
 }
